+++ src/mapping/mod.rs
@@ -5,3 +5,4 @@ pub mod meal_mapper;
 pub mod dream_mapper;
 pub mod weight_mapper;
 pub mod summary_mapper;
+pub mod error_mapper;
